Refactor to avoid circular dependency
