+++ src/discord/discord.module.ts
@@ -1,10 +1,10 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { DiscordService } from './discord.service';
 import { ConfigModule } from '@nestjs/config';
 import { SlashCommandModule } from 'src/slashCommand/slashCommand.module';
 
 @Module({
-  imports: [ConfigModule, forwardRef(() => SlashCommandModule)],
+  imports: [ConfigModule, SlashCommandModule],
   providers: [DiscordService],
   exports: [DiscordService],
 })
