+++ src/slashCommand/slashCommand.module.ts
@@ -1,17 +1,12 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { CabinModule } from 'src/cabin/cabin.module';
-import { DiscordModule } from 'src/discord/discord.module';
+import { ChatCommandModule } from 'src/chatCommand/chatCommand.module';
 import { EmbedModule } from 'src/embed/embed.module';
 import { SlashCommandService } from './slashCommand.service';
 
 @Module({
-  imports: [
-    ConfigModule,
-    CabinModule,
-    forwardRef(() => DiscordModule),
-    EmbedModule,
-  ],
+  imports: [ConfigModule, CabinModule, EmbedModule, ChatCommandModule],
   providers: [SlashCommandService],
   exports: [SlashCommandService],
 })
